Rename WorkerFactory.ts to factory.ts
